refactor(common-icons): change dir to src
https://github.com/akky-xxxx/project-my-homepage/issues/16
