+++ packages/common-icons/src/Close.tsx
